feat: Add checkToken endpoint to AuthController
This commit adds a new endpoint to the AuthController called checkToken. This endpoint is used to check the validity of a token. It returns a response indicating whether the token is valid or not.

The checkToken endpoint is implemented in the AuthController by adding a new method called checkToken. This method returns a string indicating the validity of the token.

This commit also includes changes to the AuthModule and AuthMiddleware. The AuthModule is updated to configure the AuthMiddleware for the checkToken route. The AuthMiddleware is a middleware that checks the validity of the token and throws an UnauthorizedException if the token is invalid.

These changes are necessary to provide token validation functionality in the authentication system.